RFC-VDP: Support for OUI infrastructure in VDP22.
This patch contains the header field changes for a general framework for supporting OUI fields in VDP22 at the station side. The specific changes are described below.

qbg_vdpnl.h:
-------------
Couple of parameters are added to the vdpnl_vsi structure. One is the number of OUI parameters. Generally, this will be 1, as the chances of having different OUI in a single message for VDP22 is quite low. Nevertheless, there's support for having multiple OUI in a single VSI. The other field is a general structure of len,opaque data format for carrying the OUI data.
qbg_vdp22_oui.h:
----------------
This is a new header file added for supporting the OUI framework. This file contains the OUI specific sub-structures for vdpnl and vsi22 structures and function handler structure.:w

Signed-off-by: padkrish <padkrish@cisco.com>