RFC-VDP: Support for OUI infrastructure in vdptool.
This patch contains the changes made in vdptool to support OUI fields in vdptool. This commit has only the infra-structure changes needed for supporting OUI. No specific OUI fields are added as a part of this commit. The man page for vdptool is also modified accordingly.
The OUI data can be given as input to vdptool in different ways.
It could be
 vdptool .... -c oui=companyA,Data1 -c oui=companyB,Data2 -c oui=companyA,Data3
 Or
 vdptool .... -c oui=companyA,Data1Data3 -c oui=companyB,data2
where companyA and companyB are the name of the Organizations.
Anything after the comma in OUI data field is Org specific and it's upto
the respective organization specific handlers to encode it so that it could
be decoded appropriately by the ORG specific handlers inside lldpad.
That is, Data1 and Data3 is specific to Organization 'companyA' and the
OUI handlers of companyA in vdptool and lldpad is responsible for
encoding/decoding the data. The common code in vdptool and lldpad just treats
it as opaque data. 'companyA' or 'companyB' above is the key using which
the right handlers will be called.

Irrespective of how the command line interface to vdptool is, the input to
lldpad is always the same. i.e. KeywordlenKeywordDatalenData
For OUI, the data field will have the complete OUI data starting with
ORG name (e.g companyA). So, in order to call the right handler routine,
the OUI data field is split as OUInamelenOUInameOUIData.
OUInamelen is 2B.

For example if the following is given:
vdptool -T -W -i eth2  -V assoc -c mode=assoc -c mgrid2=0 -c typeid=0 -c typeidver=0 -c uuid=18ea3452-b364-4e13-a1a2-9c6524deb685 -c hints=none -c filter=0-fa:16:3e:4c:2d:85-90001 -c oui=companyA,val1=data1

The data sent to lldpad by vdptool will be as follows assuming companyA encode
handlers in vdptool encodes it the same way:

04mode0005assoc06mgrid20001006typeid0001009typeidver0001004uuid002418ea3452-b364-4e13-a1a2-9c6524deb68505hints0004none06filter00190-fa:16:3e:4c:2d:85-9000103oui001408companyAval1=data1

This commit will not insert the oui fields as given above because no OUI
specific handlers are added.

Signed-off-by: padkrish <padkrish@cisco.com>